+++ CHANGELOG.md
@@ -17,6 +17,18 @@ CHANGELOG
 * The type of the `StatemendId::$id` property has been changed from an instance
   of `Rhumsaa\Uuid\Uuid` to an instance of the added `Uuid` class.
 
+1.2.0
+-----
+
+* The `Statement::getTimestamp()` method is deprecated and will be removed in
+  3.0. Use `Statement::getCreated()` instead.
+* The `Statement::withTimestamp()` method is deprecated and will be removed in
+  3.0. Use `Statement::withCreated()` instead.
+* The `SubStatement::getTimestamp()` method is deprecated and will be removed in
+  3.0. Use `SubStatement::getCreated()` instead.
+* The `SubStatement::withTimestamp()` method is deprecated and will be removed in
+  3.0. Use `SubStatement::withCreated()` instead.
+
 1.1.1
 -----